requirements.txt
asgiref==3.4.1
Django==3.2.14
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
Pillow==8.4.0
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2022.1
